changed int64 to int32

+++ ruby/lib/DatasetModels_pb.rb
@@ -28,7 +28,7 @@ Google::Protobuf::DescriptorPool.generated_pool.build do
       optional :Name, :string, 3
       optional :Version, :string, 4
       optional :Origin, :message, 5, "Origin"
-      optional :ContentLen, :int64, 6
+      optional :ContentLen, :int32, 6
       optional :DatasetID, :string, 7
       optional :DatasetVersionID, :string, 8
       optional :Location, :message, 9, "Location"
